container/app/requirements.txt
blinker==1.6.2
cffi==1.15.1
click==8.1.6
colorama==0.4.6
cryptography==41.0.2
Flask==2.3.2
Flask-MySQL==1.5.2
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
pycparser==2.21
PyMySQL==1.1.0
python-dotenv==1.0.0
slack-bolt==1.18.0
slack-sdk==3.21.3
Werkzeug==2.3.6
